+++ src/Render/factories/index.ts
@@ -2,3 +2,4 @@ export { createCore } from "./createCore";
 export { createInstance } from "./createInstance";
 export { registerDraws } from "./registerDraws";
 export { registerEvents } from "./registerEvents";
+export { registerTools } from "./registerTools";