+++ src/Render/interactions/index.ts
@@ -2,3 +2,4 @@ export { enableDrag } from "./drag";
 export { enableZoom } from "./zoom";
 export { enableReferenceLine } from "./referenceLine";
 export { enableDragOutside } from "./dragOutside";
+export { enableSelection } from "./selection";